refactor(editor): get loading icon with theme (#12079)
