+++ blocksuite/affine/components/src/icons/index.ts
@@ -2,6 +2,7 @@ export * from './ai.js';
 export * from './file-icons.js';
 export * from './import-export.js';
 export * from './list.js';
+export * from './loading.js';
 export * from './misc.js';
 export * from './tags.js';
 export * from './text.js';
