+++ packages/frontend/i18n/src/i18n-completenesses.json
@@ -5,7 +5,7 @@
   "de": 98,
   "el-GR": 98,
   "en": 100,
-  "es-AR": 99,
+  "es-AR": 98,
   "es-CL": 100,
   "es": 98,
   "fa": 98,
